Refatoração: Movi a função resource_path para o módulo utils.py
Movi a função utilitária `resource_path` de `src/main.py` para um novo arquivo `src/utils.py`.

Esta alteração corrige um problema de importação circular que ocorria porque `main.py` importava módulos (como `main_window.py`) que, por sua vez, precisavam importar `resource_path` de `main.py`.

Ao mover `resource_path` para um módulo de utilitários dedicado, quebrei o ciclo de dependência e melhorei a modularidade do código. Também atualizei todas as referências à função em todo o projeto para apontar para a nova localização em `src/utils.py`.
